+++ src/OpenTelemetry.Exporter.Stackdriver/StackdriverTraceExporter.cs
@@ -14,6 +14,13 @@ namespace OpenTelemetry.Exporter.Stackdriver;
 /// <summary>
 /// Exports a group of spans to Stackdriver.
 /// </summary>
+#if NET // todo: link to GCP .net documentation when becomes available
+[Obsolete("This exporter is deprecated and will be removed in a future version. Please use the native OTLP  within GCP.", UrlFormat="https://cloud.google.com/stackdriver/docs/reference/telemetry/overview")]
+#else
+[Obsolete("This exporter is deprecated and will be removed in a future version. Please use the native OTLP  within GCP.")]
+#endif
+
+
 public class StackdriverTraceExporter : BaseExporter<Activity>
 {
     private static readonly string StackdriverExportVersion;